added simple test for new story endpoint
