+++ monolith/classes/DiceSet.py
@@ -44,26 +44,6 @@ class DiceSet:
             self.pips[i] = self.dice[i].throw_die()
         return self.pips
 
-import unittest
- 
-class TestDie(unittest.TestCase):
- 
-    def test_die_init(self):
-        die = Die("tests/die0.txt")
-        check = ['bike', 'moonandstars', 'bag', 'bird', 'crying', 'angry']
-        self.assertEqual(die.faces, check)
-
-    def test_die_pip(self):
-        rnd.seed(574891)
-        die = Die("tests/die0.txt")
-        res = die.throw_die()
-        self.assertEqual(res, 'bag')
-
-    
- 
- 
-if __name__ == '__main__':
-    unittest.main()
 
 
 class InvalidDiceSet(Exception):
