+++ monolith/classes/tests/test_dice.py
@@ -1,5 +1,6 @@
 from monolith.classes.DiceSet import DiceSet, Die
 from monolith.utility.diceutils import *
+import random as rnd
 import unittest
 
 
@@ -17,6 +18,17 @@ class TestDie(unittest.TestCase):
         for i in range(0, 6):
             face_list = get_die_faces_lsit("standard", i)
             self.assertTrue(thrown[i] in face_list)
+
+    def test_die_init(self):
+        die = Die("die0.txt")
+        check = ['bike', 'moonandstars', 'bag', 'bird', 'crying', 'angry']
+        self.assertEqual(die.faces, check)
+
+    def test_die_pip(self):
+        rnd.seed(574891)
+        die = Die("die0.txt")
+        res = die.throw_die()
+        self.assertEqual(res, 'bag')
 
 
 if __name__ == '__main__':
