--- monolith/tests/test_dummy.py
@@ -1,7 +1,0 @@
-import unittest
-
-
-class TestApp(unittest.TestCase):
-
-    def test1(self):
-        print("bra")
